+++ packages/core/navigation-next/examples/0-navigation.js
@@ -81,6 +81,7 @@ const globalNavSecondaryItems = [
 const productRootNavSections = [
   {
     key: 'header',
+    isRootLevel: true,
     items: [
       {
         type: () => (
@@ -94,6 +95,7 @@ const productRootNavSections = [
   },
   {
     key: 'menu',
+    isRootLevel: true,
     items: [
       {
         type: Item,
@@ -110,6 +112,7 @@ const productRootNavSections = [
 const productContainerNavSections = [
   {
     key: 'header',
+    isRootLevel: true,
     items: [
       {
         type: ContainerHeader,
@@ -124,6 +127,7 @@ const productContainerNavSections = [
   },
   {
     key: 'menu',
+    isRootLevel: true,
     items: [
       { type: SectionTitle, key: 'title', children: 'Section title' },
       { type: Item, key: 'backlog', text: 'Backlog', before: BacklogIcon },
@@ -146,10 +150,12 @@ const GlobalNavigation = () => (
 
 const RenderSection = ({ section }: *) => (
   <div css={{ paddingTop: '16px' }}>
-    {section.map(({ key, items }) => (
+    {section.map(({ key, isRootLevel, items }) => (
       <Section key={key}>
         {({ css }) => (
-          <div css={css}>
+          <div
+            css={{ ...css, ...(isRootLevel ? { padding: '0 16px' } : null) }}
+          >
             {items.map(({ type: Component, ...props }) => (
               <Component {...props} />
             ))}