+++ packages/core/navigation-next/src/renderer/types.js
@@ -12,6 +12,7 @@ export type ItemProps = {
   before?: ComponentType<*>,
   href?: string,
   icon?: string,
+  id: string,
   goTo?: string,
 };