+++ packages/editor/editor-core/src/plugins/code-block/index.tsx
@@ -1,42 +1,70 @@
 import * as React from 'react';
+import { colors } from '@atlaskit/theme';
+import Objects24CodeIcon from '@atlaskit/icon/glyph/objects/24/code';
 import { codeBlock } from '@atlaskit/editor-common';
 import { EditorPlugin } from '../../types';
 import { plugin, stateKey } from './pm-plugins/main';
 import keymap from './pm-plugins/keymaps';
+import ideUX from './pm-plugins/ide-ux';
 import LanguagePicker from './ui/LanguagePicker';
 
-const codeBlockPlugin: EditorPlugin = {
-  nodes() {
-    return [{ name: 'codeBlock', node: codeBlock, rank: 800 }];
-  },
+export interface CodeBlockOptions {
+  enableKeybindingsForIDE?: boolean;
+}
 
-  pmPlugins() {
-    return [
-      { rank: 700, plugin: () => plugin },
-      { rank: 720, plugin: ({ schema }) => keymap(schema) },
-    ];
-  },
+const codeBlockPlugin = (options: CodeBlockOptions = {}) =>
+  ({
+    nodes() {
+      return [{ name: 'codeBlock', node: codeBlock, rank: 800 }];
+    },
 
-  contentComponent({
-    editorView,
-    appearance,
-    popupsMountPoint,
-    popupsBoundariesElement,
-  }) {
-    if (appearance === 'message') {
-      return null;
-    }
+    pmPlugins() {
+      return [
+        { rank: 700, plugin: () => plugin },
+        {
+          rank: 710,
+          plugin: () => (options.enableKeybindingsForIDE ? ideUX : undefined),
+        },
+        { rank: 720, plugin: ({ schema }) => keymap(schema) },
+      ];
+    },
 
-    const pluginState = stateKey.getState(editorView.state);
-    return (
-      <LanguagePicker
-        editorView={editorView}
-        pluginState={pluginState}
-        popupsMountPoint={popupsMountPoint}
-        popupsBoundariesElement={popupsBoundariesElement}
-      />
-    );
-  },
-};
+    contentComponent({
+      editorView,
+      appearance,
+      popupsMountPoint,
+      popupsBoundariesElement,
+    }) {
+      if (appearance === 'message') {
+        return null;
+      }
+
+      const pluginState = stateKey.getState(editorView.state);
+      return (
+        <LanguagePicker
+          editorView={editorView}
+          pluginState={pluginState}
+          popupsMountPoint={popupsMountPoint}
+          popupsBoundariesElement={popupsBoundariesElement}
+        />
+      );
+    },
+
+    pluginsOptions: {
+      quickInsert: [
+        {
+          title: 'Code block',
+          keywords: ['javascript', 'typescript'],
+          icon: () => (
+            <Objects24CodeIcon label="Code block" primaryColor={colors.N300} />
+          ),
+          action(insert, state) {
+            const schema = state.schema;
+            return insert(schema.nodes.codeBlock.createChecked());
+          },
+        },
+      ],
+    },
+  } as EditorPlugin);
 
 export default codeBlockPlugin;