+++ packages/editor/renderer/src/react/nodes/mediaSingle.tsx
@@ -38,6 +38,11 @@ export default class MediaSingle extends Component<
   { layout: MediaSingleLayout } & React.Props<any>,
   State
 > {
+  constructor(props) {
+    super(props);
+    this.state = {}; // Need to initialize with empty state.
+  }
+
   private onExternalImageLoaded = ({ width, height }) => {
     this.setState({
       width,