+++ packages/search/global-search/src/components/ObjectResult.tsx
@@ -1,10 +1,11 @@
 import * as React from 'react';
 import Avatar from '@atlaskit/avatar';
+import { colors } from '@atlaskit/theme';
 import { ResultBase } from '@atlaskit/quick-search';
 import { ResultContentType } from '../model/Result';
 import Objects24PageIcon from '@atlaskit/icon/glyph/objects/24/page';
 import Objects24BlogIcon from '@atlaskit/icon/glyph/objects/24/blog';
-import DocumentFilledIcon from '@atlaskit/icon/glyph/document-filled';
+import Objects24ImageIcon from '@atlaskit/icon/glyph/objects/24/image';
 
 const OBJECT_RESULT_TYPE = 'object';
 
@@ -29,15 +30,42 @@ export default class ObjectResult extends React.Component<Props> {
     type: OBJECT_RESULT_TYPE,
   };
 
+  /*
+   * Note:
+   * Icon with medium size = 24px.
+   * Avatar with small size = 24px.
+   * 
+   * Colors come from: https://extranet.atlassian.com/display/ADG/Object+icons%3A+Colors
+   */
   getAvatar = () => {
     if (this.props.contentType === ResultContentType.Page) {
-      return <Objects24PageIcon size="large" label={this.props.name} />;
+      return (
+        <Objects24PageIcon
+          size="medium"
+          primaryColor={colors.B200}
+          label={this.props.name}
+        />
+      );
     } else if (this.props.contentType === ResultContentType.Blogpost) {
-      return <Objects24BlogIcon size="large" label={this.props.name} />;
+      return (
+        <Objects24BlogIcon
+          size="medium"
+          primaryColor={colors.B200}
+          label={this.props.name}
+        />
+      );
     } else if (this.props.contentType === ResultContentType.Attachment) {
-      return <DocumentFilledIcon size="large" label={this.props.name} />;
+      return (
+        <Objects24ImageIcon
+          size="medium"
+          primaryColor={colors.R300}
+          label={this.props.name}
+        />
+      );
     } else {
-      return <Avatar src={this.props.avatarUrl} appearance="square" />;
+      return (
+        <Avatar src={this.props.avatarUrl} size="small" appearance="square" />
+      );
     }
   };